rwanda-bus-api/requirements.txt
anyio==4.9.0
bcrypt==4.0.1
click==8.1.8
dnspython==2.3.0
ecdsa==0.19.1
email-validator==2.0.0
fastapi==0.95.1
h11==0.14.0
idna==3.10
loguru==0.7.0
motor==3.1.2
passlib==1.7.4
pyasn1==0.6.1
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.26.1
typing-extensions==4.13.2
uvicorn==0.22.0
